backend/src/requirements.txt
fastapi
mangum
boto3
sqlalchemy
psycopg2-binary
uvicorn
pydantic
passlib[bcrypt]
python-jose[cryptography]